Refactor of my hl7-example
I've done a great deal of refactoring. I've refactored the db part to take in the e0 (first element) of the HL7 record, and I've refactored the Active Record code to match.  As well, I've changed the looping to be much more dynamic based on the number of fields it finds values for and places them into the matching field in the db.